services/detector-megadetector/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pillow==10.4.0
httpx==0.27.0
megadetector
